+++ markdown-mode.el
@@ -192,45 +192,43 @@
 ;;
 ;;   * Hyperlinks: `C-c C-a`
 ;;
-;;     In this group, `C-c C-a l` inserts an inline link of the form
-;;     `[text](url)`.  The link text is determined as follows.  First,
-;;     if there is an active region (i.e., when transient mark mode is
-;;     on and the mark is active), use it as the link text.  Second,
-;;     if the point is at a word, use that word as the link text.  In
-;;     these two cases, the original text will be replaced with the
-;;     link and point will be left at the position for inserting a
-;;     URL.  Otherwise, insert empty link markup and place the point
-;;     for inserting the link text.
+;;     In this group, `C-c C-a l` inserts links, either inline,
+;;     reference, or plain URLs.  The URL or `[reference]` label, link
+;;     text, and optional title are entered through a series of
+;;     interactive prompts.  The type of link is determined by which
+;;     values are provided:
 ;;
-;;     `C-c C-a L` inserts a reference link of the form `[text][label]`
-;;     and, optionally, a corresponding reference label definition.
-;;     The link text is determined in the same way as with an inline
-;;     link (using the region, when active, or the word at the point),
-;;     but instead of inserting empty markup as a last resort, the
-;;     link text will be read from the minibuffer.  The reference
-;;     label will be read from the minibuffer in both cases, with
-;;     completion from the set of currently defined references.  To
-;;     create an implicit reference link, press `RET` to accept the
-;;     default, an empty label.  If the entered referenced label is
-;;     not defined, additionally prompt for the URL and (optional)
-;;     title.  If a URL is provided, a reference definition will be
-;;     inserted in accordance with `markdown-reference-location'.
-;;     If a title is given, it will be added to the end of the
-;;     reference definition and will be used to populate the title
-;;     attribute when converted to XHTML.
+;;     *   If both a URL and link text are given, insert an inline link:
+;;         `[text](url)`.
+;;     *   If both a `[reference]` label and link text are given, insert
+;;         a reference link: `[text][reference]`.
+;;     *   If only link text is given, insert an implicit reference link:
+;;         `[text][]`.
+;;     *   If only a URL is given, insert a plain URL link:
+;;         `<url>`.
 ;;
-;;     `C-c C-a u` inserts a bare url, delimited by angle brackets.  When
-;;     there is an active region, the text in the region is used as the
-;;     URL.  If the point is at a URL, that url is used.  Otherwise,
-;;     insert angle brackets and position the point in between them
-;;     for inserting the URL.
+;;     If there is an active region, use the text as the default URL,
+;;     if it seems to be a URL, or link text value otherwise.  The region
+;;     will be deleted and replaced by the link.
+;;
+;;     Note that this function can be used to convert a link from one
+;;     type to another (inline, reference, or plain URL) by
+;;     selectively adding or removing properties via the interactive
+;;     prompts.
+;;
+;;     If a reference label is given that is not yet defined, you
+;;     will be prompted for the URL and optional title and the
+;;     reference will be inserted according to the value of
+;;     `markdown-reference-location'.  If a title is given, it will be
+;;     added to the end of the reference definition and will be used
+;;     to populate the title attribute when converted to HTML.
 ;;
 ;;     `C-c C-a f` inserts a footnote marker at the point, inserts a
 ;;     footnote definition below, and positions the point for
 ;;     inserting the footnote text.  Note that footnotes are an
 ;;     extension to Markdown and are not supported by all processors.
 ;;
-;;     `C-c C-a w` behaves much like the inline link insertion command
+;;     `C-c C-a w` behaves much like the link insertion command
 ;;     and inserts a wiki link of the form `[[WikiLink]]`.  If there
 ;;     is an active region, use the region as the link text.  If the
 ;;     point is at a word, use the word as the link text.  If there is
@@ -4162,56 +4160,6 @@ Optionally, the user can provide a TITLE."
     (cond ((not text) (goto-char (+ 1 cur)))
           ((not url) (goto-char (+ 3 (length text) cur))))))
 
-(defun markdown-insert-inline-link-dwim ()
-  "Insert an inline link of the form [link](url) at point.
-If there is an active region, the text in the region will be used
-as the URL, if it appears to be a URL, or else as the link text.
-If the point is at a URL, use it to create a new link. If the
-point is at a reference link, convert it to an inline link. If
-the point is at a word, use the word as the link text. In these
-cases, the point will be left at the position for inserting a
-URL. If there is no active region and the point is not at word,
-simply insert link markup and place the point in the position to
-enter link text."
-  (interactive)
-  (cond
-   ;; If there is an active region, remove existing links in the
-   ;; region and use resulting region as link text for a new link.
-   ((markdown-use-region-p)
-    (let* ((bounds (markdown-unwrap-things-in-region
-                    (region-beginning) (region-end)
-                    markdown-regex-link-inline 0 3))
-           (text (buffer-substring (car bounds) (cdr bounds))))
-      (delete-region (car bounds) (cdr bounds))
-      (markdown-insert-inline-link text nil)))
-   ;; If there is an inline link at the point, remove it and add the
-   ;; link text to the kill ring.
-   ((thing-at-point-looking-at markdown-regex-link-inline)
-    (kill-new (match-string 3))
-    (delete-region (match-beginning 0) (match-end 0)))
-   ;; If there is an angle URL at the point, use it for a new link.
-   ((thing-at-point-looking-at markdown-regex-angle-uri)
-    (let ((url (match-string 2)))
-      (delete-region (match-beginning 0) (match-end 0))
-      (markdown-insert-inline-link nil url)))
-   ;; If there is a plain URL at the point, use it for a new link.
-   ((thing-at-point-looking-at markdown-regex-uri)
-    (let ((url (match-string 0)))
-      (delete-region (match-beginning 0) (match-end 0))
-      (markdown-insert-inline-link nil url)))
-   ;; If there is a reference link at point, convert to inline link.
-   ((thing-at-point-looking-at markdown-regex-link-reference)
-    (let ((beg (match-beginning 0))
-          (end (match-end 0))
-          (text (match-string 3))
-          (url (markdown-link-url)))
-      (delete-region beg end)
-      (markdown-insert-inline-link text url)))
-   ;; Otherwise, insert a link
-   (t (let ((bounds (markdown-wrap-or-insert "[" "]()")))
-        (when bounds
-          (goto-char (- (cdr bounds) 1)))))))
-
 (defun markdown-insert-reference-link (text label &optional url title)
   "Insert a reference link and, optionally, a reference definition.
 The link TEXT will be inserted followed by the optional LABEL.
@@ -4258,48 +4206,90 @@ be used to populate the title attribute when converted to XHTML."
         "Defined reference [%s], press \\[markdown-jump] to jump there")
        label))))
 
-(defun markdown-insert-reference-link-dwim ()
-  "Insert a reference link of the form [text][label] at point.
-If there is an active region, the text in the region will be used
-as the link text.  If the point is at an inline link, it will be
-converted to a reference link.  If the point is at a word, it will
-be used as the link text.  Otherwise, the link text will be read from
-the minibuffer.  The link label will be read from the minibuffer in
-both cases, with completion from the set of currently defined
-references.  To create an implicit reference link, press RET to
-accept the default, an empty label.  If the entered referenced
-label is not defined, additionally prompt for the URL
-and (optional) title.  The reference definition is placed at the
-location determined by `markdown-reference-location'."
+(make-obsolete 'markdown-insert-inline-link-dwim 'markdown-insert-link "v2.3")
+(make-obsolete 'markdown-insert-reference-link-dwim 'markdown-insert-link "v2.3")
+
+(defun markdown-insert-link ()
+  "Insert new or update an existing link, with interactive prompts.
+If the point is at an existing link or URL, update the link text,
+URL, reference label, and/or title.  Otherwise, insert a new link.
+The type of link inserted (inline, reference, or plain URL)
+depends on which values are provided:
+
+*   If a URL and TEXT are given, insert an inline link: [TEXT](URL).
+*   If [REF] and TEXT are given, insert a reference link: [TEXT][REF].
+*   If only TEXT is given, insert an implicit reference link: [TEXT][].
+*   If only a URL is given, insert a plain link: <URL>.
+
+In other words, to create an implicit reference link, leave the
+URL prompt empty and to create a plain URL link, leave the link
+text empty.
+
+If there is an active region, use the text as the default URL, if
+it seems to be a URL, or link text value otherwise.
+
+If a given reference is not defined, this function will
+additionally prompt for the URL and optional title.  In this case,
+the reference definition is placed at the location determined by
+`markdown-reference-location'.
+
+Through updating the link, this function can be used to convert a
+link of one type (inline, reference, or plain) to another type by
+selectively adding or removing information via the prompts."
   (interactive)
-  (let* ((defined-labels (markdown-get-defined-references))
-         (switch (thing-at-point-looking-at markdown-regex-link-inline))
-         (bounds (cond ((markdown-use-region-p)
-                        (cons (region-beginning) (region-end)))
-                       (switch
-                        (cons (match-beginning 0) (match-end 0)))
-                       (t
-                        (markdown-bounds-of-thing-at-point 'word))))
-         (text (cond (switch (match-string 3))
-                     (bounds (buffer-substring (car bounds) (cdr bounds)))
-                     (t (read-string "Link Text: "))))
-         (label (completing-read
-                 "Link Label (default: none): " defined-labels
-                 nil nil nil 'markdown-reference-label-history nil))
-         (ref (save-match-data
-                (markdown-reference-definition
-                 (if (> (length label) 0) label text))))
-         (url (cond (ref nil)
-                    (switch (match-string 6))
-                    (t (read-string "Link URL: "))))
-         (title (cond
-                 ((= (length url) 0) nil)
-                 (switch (if (> (length (match-string 7)) 2)
-                             (substring (match-string 7) 1 -1)
-                           nil))
-                 (t (read-string "Link Title (optional): ")))))
-    (when bounds (delete-region (car bounds) (cdr bounds)))
-    (markdown-insert-reference-link text label url title)))
+  (cl-multiple-value-bind (begin end text uri ref title)
+      (if (markdown-use-region-p)
+          ;; Use region as either link text or URL as appropriate.
+          (let ((region (buffer-substring-no-properties
+                         (region-beginning) (region-end))))
+            (if (string-match markdown-regex-uri region)
+                ;; Region contains a URL; use it as such.
+                (list (region-beginning) (region-end)
+                      nil (match-string 0 region) nil nil)
+              ;; Region doesn't contain a URL, so use it as text.
+              (list (region-beginning) (region-end)
+                    region nil nil nil)))
+        ;; Extract and use properties of existing link, if any.
+        (markdown-link-at-pos (point)))
+    (let* ((ref (when ref (concat "[" ref "]")))
+           (defined-refs (append
+                          (mapcar (lambda (ref) (concat "[" ref "]"))
+                                  (markdown-get-defined-references))))
+           (used-uris (markdown-get-used-uris))
+           (uri-or-ref (completing-read
+                        "URI or [reference]: "
+                        (append defined-refs used-uris)
+                        nil nil (or uri ref)))
+           (ref (cond ((string-match "\\`\\[\\(.*\\)\\]\\'" uri-or-ref)
+                       (match-string 1 uri-or-ref))
+                      ((string-equal "" uri-or-ref)
+                       "")))
+           (uri (unless ref uri-or-ref))
+           (text-prompt (if ref
+                            "Link text: "
+                          "Link text (blank for plain URI): "))
+           (text (read-string text-prompt text))
+           (text (if (= (length text) 0) nil text))
+           (plainp (and uri (not text)))
+           (implicitp (string-equal ref ""))
+           (ref (if implicitp text ref))
+           (definedp (and ref (markdown-reference-definition ref)))
+           (ref-url (unless (or uri definedp)
+                      (completing-read "Reference URI: " used-uris)))
+           (title (unless (or plainp definedp)
+                    (read-string "Title or alt text (optional): " title)))
+           (title (if (= (length title) 0) nil title)))
+      (when (and begin end)
+        (delete-region begin end))
+      (cond
+       ((and uri text)
+        (markdown-insert-inline-link text uri title))
+       ((and ref text)
+        (markdown-insert-reference-link text (unless implicitp ref) nil title)
+        (unless definedp
+          (markdown-insert-reference-definition ref ref-url title)))
+       (uri
+        (markdown-insert-uri uri))))))
 
 (defun markdown-insert-uri (&optional uri)
   "Insert markup for an inline URI.
@@ -5471,9 +5461,7 @@ Assumes match data is available for `markdown-regex-italic'."
 (defvar markdown-mode-map
   (let ((map (make-keymap)))
     ;; Element insertion
-    (define-key map "\C-c\C-al" 'markdown-insert-inline-link-dwim)
-    (define-key map "\C-c\C-aL" 'markdown-insert-reference-link-dwim)
-    (define-key map "\C-c\C-au" 'markdown-insert-uri)
+    (define-key map "\C-c\C-al" 'markdown-insert-link)
     (define-key map "\C-c\C-af" 'markdown-insert-footnote)
     (define-key map "\C-c\C-aw" 'markdown-insert-wiki-link)
     (define-key map "\C-c\C-ii" 'markdown-insert-image)
@@ -5502,6 +5490,8 @@ Assumes match data is available for `markdown-regex-italic'."
     (define-key map "\C-c\C-ar" 'markdown-insert-reference-link-dwim)
     (define-key map "\C-c\C-tt" 'markdown-insert-header-setext-1)
     (define-key map "\C-c\C-ts" 'markdown-insert-header-setext-2)
+    (define-key map "\C-c\C-aL" 'markdown-insert-link)
+    (define-key map "\C-c\C-au" 'markdown-insert-link)
     ;; Element removal
     (define-key map (kbd "C-c C-k") 'markdown-kill-thing-at-point)
     ;; Promotion, Demotion, Completion, and Cycling
@@ -5676,13 +5666,11 @@ See also `markdown-mode-map'.")
      ["Renumber List" markdown-cleanup-list-numbers]
      ["Toggle Task List Item" markdown-toggle-gfm-checkbox])
     ("Links & Images"
-     ["Plain URL" markdown-insert-uri]
-     ["Inline Link" markdown-insert-inline-link-dwim]
-     ["Inline Image" markdown-insert-image]
-     ["Reference Link" markdown-insert-reference-link-dwim]
-     ["Reference Image" markdown-insert-reference-image]
-     ["Footnote" markdown-insert-footnote]
-     ["Wiki Link" markdown-insert-wiki-link]
+     ["Insert Link" markdown-insert-link]
+     ["Insert Inline Image" markdown-insert-image]
+     ["Insert Reference Image" markdown-insert-reference-image]
+     ["Insert Footnote" markdown-insert-footnote]
+     ["Insert Wiki Link" markdown-insert-wiki-link]
      "---"
      ["Check References" markdown-check-refs]
      ["Toggle URL Hiding" markdown-toggle-url-hiding